# static/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "0.1.0",
    "title": "Go-Fast CDN",
    "description": "A simple and fast CDN written in Go",
    "contact": {
      "name": "Jonas Lateur",
      "email": "jonas@lateur.pro",
      "url": "https://git.buildserver.be/jayjay1989be/go-fast-cdn"
    },
    "license": {
      "name": "MIT License",
      "url": "https://git.buildserver.be/jayjay1989be/go-fast-cdn/raw/main/LICENSE"
    }
  },
  "paths": {
...
    "/api/cdn/delete/file/{filename}": {
      "delete": {
        "summary": "Delete file",
        "description": "Delete an document from the CDN.",
        "operationId": "deleteFile",
        "parameters": [
          {
            "name": "filename",
            "in": "path",
            "description": "The filename of the file to be deleted.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "You have deleted the requested file."
          },
          "400": {
            "description": "Missing or invalid parameters"
          }
        }
      }
    },
...
    "/api/cdn/rename/file": {
      "put": {
        "summary": "Rename document",
        "description": "Rename an file from the CDN.",
        "operationId": "renameFile",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "filename": {
                    "type": "string",
                    "description": "The filename you want to rename"
                  },
                  "newname": {
                    "type": "string",
                    "description": "The name you want the file to have"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "You have renamed the requested file."
          },
          "400": {
            "description": "Missing or invalid body"
          }
        }
      }
    },
...
  },
...
}